+++ js_modules/dagit/packages/core/src/runs/RunTable.tsx
@@ -7,8 +7,7 @@ import styled from 'styled-components/macro';
 import {isHiddenAssetGroupJob} from '../asset-graph/Utils';
 import {RunsFilter} from '../graphql/types';
 import {useSelectionReducer} from '../hooks/useSelectionReducer';
-import {useLaunchPadHooks} from '../launchpad/LaunchpadHooksContext';
-import {PipelineSnapshotLink, getPipelineSnapshotLink} from '../pipelines/PipelinePathUtils';
+import {PipelineSnapshotLink} from '../pipelines/PipelinePathUtils';
 import {PipelineReference} from '../pipelines/PipelineReference';
 import {AnchorButton} from '../ui/AnchorButton';
 import {
@@ -23,7 +22,6 @@ import {workspacePipelinePath, workspacePipelinePathGuessRepo} from '../workspac
 import {AssetKeyTagCollection} from './AssetKeyTagCollection';
 import {RunActionsMenu, RunBulkActionsMenu} from './RunActionsMenu';
 import {RunStatusTagWithStats} from './RunStatusTag';
-import {DagsterTag, TagType} from './RunTag';
 import {RunTags} from './RunTags';
 import {
   assetKeysForRun,
@@ -141,12 +139,11 @@ export const RunTable = (props: RunTableProps) => {
                 />
               ) : null}
             </th>
-            <th style={{width: 90}}>Run ID</th>
-            <th style={{width: 180}}>Created Date</th>
-            <th>Target</th>
-            <th style={{width: 160}}>Created By</th>
             <th style={{width: 120}}>Status</th>
-            <th style={{width: 190}}>Duration</th>
+            <th style={{width: 90}}>Run ID</th>
+            <th>{anyPipelines ? 'Job / Pipeline' : 'Job'}</th>
+            <th style={{width: 90}}>Snapshot ID</th>
+            <th style={{width: 190}}>Timing</th>
             {props.additionalColumnHeaders}
             <th style={{width: 52}} />
           </tr>
@@ -244,8 +241,6 @@ const RunRow: React.FC<{
     }
   };
 
-  const {RunCreatedByCell} = useLaunchPadHooks();
-
   return (
     <Row highlighted={!!isHighlighted}>
       <td>
@@ -254,12 +249,12 @@ const RunRow: React.FC<{
         ) : null}
       </td>
       <td>
+        <RunStatusTagWithStats status={run.status} runId={run.id} />
+      </td>
+      <td>
         <Link to={`/runs/${run.id}`}>
           <Mono>{titleForRun(run)}</Mono>
         </Link>
-      </td>
-      <td>
-        <RunTime run={run} />
       </td>
       <td>
         <Box flex={{direction: 'column', gap: 5}}>
@@ -291,35 +286,26 @@ const RunRow: React.FC<{
             </Box>
           )}
           <RunTags
-            tags={
-              [
-                {
-                  key: DagsterTag.SnapshotID,
-                  value: run.pipelineSnapshotId?.slice(0, 8) || '',
-                  link: run.pipelineSnapshotId
-                    ? getPipelineSnapshotLink(run.pipelineName, run.pipelineSnapshotId)
-                    : undefined,
-                },
-                run.tags.find((tag) => tag.key === DagsterTag.Partition),
-              ].filter((x) => x) as TagType[]
-            }
+            tags={run.tags}
             mode={isJob ? (run.mode !== 'default' ? run.mode : null) : run.mode}
             onAddTag={onAddTag}
           />
         </Box>
       </td>
       <td>
-        <RunCreatedByCell run={run} />
+        <PipelineSnapshotLink
+          snapshotId={run.pipelineSnapshotId || ''}
+          pipelineName={run.pipelineName}
+          size="normal"
+        />
       </td>
       <td>
-        <RunStatusTagWithStats status={run.status} runId={run.id} />
-      </td>
-      <td>
+        <RunTime run={run} />
         <RunStateSummary run={run} />
       </td>
       {additionalColumns}
       <td>
-        <RunActionsMenu run={run} onAddTag={onAddTag} />
+        <RunActionsMenu run={run} />
       </td>
     </Row>
   );
