+++ js_modules/dagit/packages/core/src/runs/RunTableNew.tsx
@@ -7,7 +7,8 @@ import styled from 'styled-components/macro';
 import {isHiddenAssetGroupJob} from '../asset-graph/Utils';
 import {RunsFilter} from '../graphql/types';
 import {useSelectionReducer} from '../hooks/useSelectionReducer';
-import {PipelineSnapshotLink} from '../pipelines/PipelinePathUtils';
+import {useLaunchPadHooks} from '../launchpad/LaunchpadHooksContext';
+import {PipelineSnapshotLink, getPipelineSnapshotLink} from '../pipelines/PipelinePathUtils';
 import {PipelineReference} from '../pipelines/PipelineReference';
 import {AnchorButton} from '../ui/AnchorButton';
 import {
@@ -22,6 +23,7 @@ import {workspacePipelinePath, workspacePipelinePathGuessRepo} from '../workspac
 import {AssetKeyTagCollection} from './AssetKeyTagCollection';
 import {RunActionsMenu, RunBulkActionsMenu} from './RunActionsMenu';
 import {RunStatusTagWithStats} from './RunStatusTag';
+import {DagsterTag, TagType} from './RunTag';
 import {RunTags} from './RunTags';
 import {
   assetKeysForRun,
@@ -139,11 +141,12 @@ export const RunTable = (props: RunTableProps) => {
                 />
               ) : null}
             </th>
-            <th style={{width: 120}}>Status</th>
             <th style={{width: 90}}>Run ID</th>
-            <th>{anyPipelines ? 'Job / Pipeline' : 'Job'}</th>
-            <th style={{width: 90}}>Snapshot ID</th>
-            <th style={{width: 190}}>Timing</th>
+            <th style={{width: 180}}>Created Date</th>
+            <th>Target</th>
+            <th style={{width: 160}}>Created By</th>
+            <th style={{width: 120}}>Status</th>
+            <th style={{width: 190}}>Duration</th>
             {props.additionalColumnHeaders}
             <th style={{width: 52}} />
           </tr>
@@ -241,6 +244,8 @@ const RunRow: React.FC<{
     }
   };
 
+  const {RunCreatedByCell} = useLaunchPadHooks();
+
   return (
     <Row highlighted={!!isHighlighted}>
       <td>
@@ -249,12 +254,12 @@ const RunRow: React.FC<{
         ) : null}
       </td>
       <td>
-        <RunStatusTagWithStats status={run.status} runId={run.id} />
-      </td>
-      <td>
         <Link to={`/runs/${run.id}`}>
           <Mono>{titleForRun(run)}</Mono>
         </Link>
+      </td>
+      <td>
+        <RunTime run={run} />
       </td>
       <td>
         <Box flex={{direction: 'column', gap: 5}}>
@@ -286,26 +291,35 @@ const RunRow: React.FC<{
             </Box>
           )}
           <RunTags
-            tags={run.tags}
+            tags={
+              [
+                {
+                  key: DagsterTag.SnapshotID,
+                  value: run.pipelineSnapshotId?.slice(0, 8) || '',
+                  link: run.pipelineSnapshotId
+                    ? getPipelineSnapshotLink(run.pipelineName, run.pipelineSnapshotId)
+                    : undefined,
+                },
+                run.tags.find((tag) => tag.key === DagsterTag.Partition),
+              ].filter((x) => x) as TagType[]
+            }
             mode={isJob ? (run.mode !== 'default' ? run.mode : null) : run.mode}
             onAddTag={onAddTag}
           />
         </Box>
       </td>
       <td>
-        <PipelineSnapshotLink
-          snapshotId={run.pipelineSnapshotId || ''}
-          pipelineName={run.pipelineName}
-          size="normal"
-        />
+        <RunCreatedByCell run={run} />
       </td>
       <td>
-        <RunTime run={run} />
+        <RunStatusTagWithStats status={run.status} runId={run.id} />
+      </td>
+      <td>
         <RunStateSummary run={run} />
       </td>
       {additionalColumns}
       <td>
-        <RunActionsMenu run={run} />
+        <RunActionsMenu run={run} onAddTag={onAddTag} />
       </td>
     </Row>
   );
